backend/requirements.txt
flask
flask-cors
pymongo
python-dotenv
tensorflow
opencv-python
numpy
ultralytics
pillow
gunicorn
requests
